projects/conditional-residual/pyproject.toml
[tool.poetry]
name = "conditional-residual"
version = "0.0.1rc0"
description = "Conditional and residual methods in scalable coding for humans and machines."
authors = ["Anderson de Andrade <anderson_de_andrade@sfu.ca>"]
include = ["coding/*"]
exclude = ["tests/*"]

[[tool.poetry.source]]
name = "private"
url = "https://pypi.org/simple"
secondary = true

[tool.poetry.dependencies]
python = "3.10.6"
annoy = "1.17.2"
compressai = "1.2.4"
defopt = "6.4.0"
matplotlib = "*"
matplotlib-venn = "0.11.9"
mlflow = "2.3.1"
numba = "0.56.4"
numpy = "*"
pillow = "*"
pycocotools = "2.0.6"
pytorch-lightning = "1.9.3"
scipy = "1.10.1"
seaborn = "0.12.2"
sfu-torch-lib = "0.0.38"
torch = "1.13.1"

[tool.poetry.extras]
extras = []

[tool.poetry.dev-dependencies]
defopt = "*"
fabric = "*"
flake8 = "*"
mlflow-skinny = "*"
mypy = "*"
pytest = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
